Intelligent Turtle
This code is an extension of turtle graphics, in the previous turtle it makes lines but it cannot repeat its task. This code provides ability to remember where the turtle was previously be saving the coordinates into the vector list. In this way we can repeat the drawing for endless times.